+++ packages/polaris-viz/src/components/Annotations/constants.ts
@@ -1,6 +1,7 @@
 export const PILL_HEIGHT = 20;
-export const PILL_PADDING = 10;
+export const PILL_PADDING = 8;
 export const PILL_ROW_GAP = 8;
 export const PILL_X_MIN = -10;
-export const SHOW_MORE_BUTTON_OFFSET = 27;
 export const ANNOTATION_Y_AXIS_LABEL_HEIGHT = 11.5;
+export const CONTENT_Y_OFFSET = PILL_HEIGHT + PILL_ROW_GAP;
+export const LABEL_MOUSEOFF_DELAY = 200;