+++ packages/polaris-viz/src/components/HorizontalBarChart/components/HorizontalBarChartXAnnotations/HorizontalBarChartXAnnotations.tsx
@@ -1,8 +1,13 @@
 import type {ReactNode} from 'react';
-import {Fragment, useMemo, useState} from 'react';
+import {Fragment, useMemo, useRef, useState} from 'react';
 import type {ScaleLinear} from 'd3-scale';
 import {isValueWithinDomain} from '@shopify/polaris-viz-core';
 
+// eslint-disable-next-line @shopify/strict-component-boundaries
+import {
+  CONTENT_Y_OFFSET,
+  LABEL_MOUSEOFF_DELAY,
+} from '../../../Annotations/constants';
 import type {
   Annotation,
   AnnotationLookupTable,
@@ -13,12 +18,8 @@ import {
   AnnotationLabel,
   AnnotationLine,
   AnnotationContent,
-  ShowMoreAnnotationsButton,
   PILL_HEIGHT,
-  SHOW_MORE_BUTTON_OFFSET,
 } from '../../../Annotations';
-import {isShowMoreAnnotationsButtonVisible} from '../../../../utilities/isShowMoreAnnotationsButtonVisible';
-import {shouldHideAnnotation} from '../../../../utilities/shouldHideAnnotation';
 
 import {useHorizontalBarChartXAnnotationPositions} from './hooks/useHorizontalBarChartXAnnotationPositions';
 
@@ -40,8 +41,8 @@ export function HorizontalBarChartXAnnotations({
   xScale,
 }: AnnotationsProps) {
   const [activeIndex, setActiveIndex] = useState(-1);
-  const [isShowingAllAnnotations, setIsShowingAllAnnotations] = useState(false);
   const [ref, setRef] = useState<SVGGElement | null>(null);
+  const timeoutRef = useRef<number>(0);
 
   const {annotations} = useMemo(() => {
     const annotations = Object.keys(annotationsLookupTable)
@@ -69,22 +70,30 @@ export function HorizontalBarChartXAnnotations({
     return {annotations};
   }, [annotationsLookupTable, xScale]);
 
-  const {hiddenAnnotationsCount, positions, rowCount} =
-    useHorizontalBarChartXAnnotationPositions({
-      annotations,
-      drawableWidth,
-      isShowingAllAnnotations,
-      onHeightChange,
-      xScale,
-    });
-
-  const handleToggleAllAnnotations = () => {
-    setIsShowingAllAnnotations(!isShowingAllAnnotations);
-  };
+  const positions = useHorizontalBarChartXAnnotationPositions({
+    annotations,
+    drawableWidth,
+    onHeightChange,
+    xScale,
+  });
 
   const handleOnMouseLeave = () => {
     setActiveIndex(-1);
   };
+
+  function handleLabelMouseEnter(index: number) {
+    setActiveIndex(index);
+  }
+
+  const handleLabelMouseLeave = () => {
+    timeoutRef.current = window.setTimeout(() => {
+      setActiveIndex(-1);
+    }, LABEL_MOUSEOFF_DELAY);
+  };
+
+  function handleContentMouseEnter() {
+    clearTimeout(timeoutRef.current);
+  }
 
   useSVGBlurEvent({
     ref,
@@ -96,73 +105,51 @@ export function HorizontalBarChartXAnnotations({
     },
   });
 
-  const isShowMoreButtonVisible = isShowMoreAnnotationsButtonVisible(rowCount);
-  const showMoreButtonOffset = isShowMoreButtonVisible
-    ? SHOW_MORE_BUTTON_OFFSET
-    : 0;
-
   return (
     <g ref={setRef} tabIndex={-1}>
-      {isShowMoreButtonVisible && (
-        <ShowMoreAnnotationsButton
-          annotationsCount={hiddenAnnotationsCount}
-          collapseText={annotations[0].collapseButtonText}
-          expandText={annotations[0].expandButtonText}
-          isShowingAllAnnotations={isShowingAllAnnotations}
-          onClick={handleToggleAllAnnotations}
-          tabIndex={annotations.length}
-          width={drawableWidth}
-        />
-      )}
-      <g transform={`translate(0, ${showMoreButtonOffset})`}>
-        {positions.map((position) => {
-          const {line, y, row, index} = position;
-          const annotation = annotations[index];
+      {positions.map((position) => {
+        const {line, y, index} = position;
+        const annotation = annotations[index];
 
-          if (shouldHideAnnotation({row, isShowingAllAnnotations, rowCount})) {
-            return null;
-          }
+        const hasContent = annotation.content != null;
+        const isContentVisible = index === activeIndex && hasContent;
+        const tabIndex = index + 1;
+        const ariaLabel = `${annotation.startKey}`;
 
-          const hasContent = annotation.content != null;
-          const isContentVisible = index === activeIndex && hasContent;
-          const tabIndex = index + 1;
-          const ariaLabel = `${annotation.startKey}`;
-
-          return (
-            <Fragment key={`annotation${index}${annotation.startKey}`}>
-              <AnnotationLine
-                size={drawableHeight - showMoreButtonOffset}
-                x={line.x}
-                y={y + PILL_HEIGHT}
-              />
-              <AnnotationLabel
-                ariaLabel={ariaLabel}
-                hasContent={hasContent}
+        return (
+          <Fragment key={`annotation${index}${annotation.startKey}`}>
+            <AnnotationLine
+              size={drawableHeight}
+              x={line.x}
+              y={y + PILL_HEIGHT}
+            />
+            <AnnotationLabel
+              ariaLabel={ariaLabel}
+              index={index}
+              label={annotation.label}
+              onMouseEnter={handleLabelMouseEnter}
+              onMouseLeave={handleLabelMouseLeave}
+              position={position}
+              tabIndex={tabIndex}
+            />
+            {isContentVisible && (
+              <AnnotationContent
+                annotation={annotation}
+                drawableWidth={drawableWidth}
                 index={index}
-                isVisible={!isContentVisible}
-                label={annotation.label}
+                onMouseEnter={handleContentMouseEnter}
+                onMouseLeave={handleLabelMouseLeave}
+                parentRef={ref}
                 position={position}
-                setActiveIndex={setActiveIndex}
+                renderAnnotationContent={renderAnnotationContent}
                 tabIndex={tabIndex}
+                x={line.x}
+                y={y + CONTENT_Y_OFFSET}
               />
-              {isContentVisible && (
-                <AnnotationContent
-                  annotation={annotation}
-                  drawableWidth={drawableWidth}
-                  index={index}
-                  onMouseLeave={handleOnMouseLeave}
-                  parentRef={ref}
-                  position={position}
-                  renderAnnotationContent={renderAnnotationContent}
-                  tabIndex={tabIndex}
-                  x={line.x}
-                  y={y}
-                />
-              )}
-            </Fragment>
-          );
-        })}
-      </g>
+            )}
+          </Fragment>
+        );
+      })}
     </g>
   );
 }