+++ packages/polaris-viz/src/components/VerticalBarChart/Chart.tsx
@@ -306,6 +306,21 @@ export function Chart({
           />
         </g>
 
+        {hasYAxisAnnotations && (
+          <g
+            transform={`translate(${chartXPosition},${chartYPosition})`}
+            tabIndex={-1}
+          >
+            <YAxisAnnotations
+              annotationsLookupTable={annotationsLookupTable}
+              drawableHeight={annotationsDrawableHeight}
+              drawableWidth={drawableWidth}
+              ticks={ticks}
+              yScale={yScale}
+            />
+          </g>
+        )}
+
         {hasXAxisAnnotations && (
           <g
             transform={`translate(${chartXPosition + xAxisLabelHalf},0)`}
@@ -321,21 +336,6 @@ export function Chart({
               onHeightChange={setAnnotationsHeight}
               xScale={xScale}
               renderAnnotationContent={renderAnnotationContent}
-            />
-          </g>
-        )}
-
-        {hasYAxisAnnotations && (
-          <g
-            transform={`translate(${chartXPosition},${chartYPosition})`}
-            tabIndex={-1}
-          >
-            <YAxisAnnotations
-              annotationsLookupTable={annotationsLookupTable}
-              drawableHeight={annotationsDrawableHeight}
-              drawableWidth={drawableWidth}
-              ticks={ticks}
-              yScale={yScale}
             />
           </g>
         )}