+++ packages/polaris-viz/CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Removed
+
+- Removed `Show more annotations` button. We now render all available annotations.
+
+### Changed
+
+- Tweaks annotation pill visuals.
 
 ## [16.14.0] - 2025-04-25